fix: update docs
