+++ docs/OPTIONS.md
@@ -1,16 +1,16 @@
 ## Plugin options
 
-| Option                |Type      | Required | Default value  | Description                                                           |
-| --------------------- |----------|----------|----------------|-----------------------------------------------------------------------|
-| `outputDir`           |`string`  |false     | regression     | Working directory. Images, report, etc will be saved into this folder |
-| `instanceFolder`      |`Function`|false     | undefined      | Callback to create folder by specific browser. See details below      |
-| `customMatchers`      |`string[]`|false     | [ ]            | Register your custom matchers to include it for output report         |
-| `largeImageThreshold` |`number`  |false     | 1200           | Skips pixels when the image width or height is larger than this one   |
-| `allowedMismatch`     |`number`  |false     | 0.1            | Ignore failed result if mismatch percentage less than allowed         |
+| Option                |Type        | Required | Default value  | Description                                                           |
+| --------------------- |------------|----------|----------------|-----------------------------------------------------------------------|
+| `outputDir`           |`string`    |false     | regression     | Working directory. Images, report, etc will be saved into this folder |
+| `instanceFolder`      |`Function`  |false     | undefined      | Callback to create folder by specific browser. See details below      |
+| `customMatchers`      |`Function[]`|false     | [ ]            | Add custom matchers to webdriverio                                    |
+| `largeImageThreshold` |`number`    |false     | 1200           | Skips pixels when the image width or height is larger than this one   |
+| `allowedMismatch`     |`number`    |false     | 0.1            | Ignore failed result if mismatch percentage less than allowed         |
 
 #### instanceFolder
 
-This option allow to provide custom folder name for each browser. The callback must return folder name as string. For example:
+This option allow to provide custom folder name per each browser. The callback must return folder name as string. For example:
 
 ```ts
 instanceFolder: info => {
