+++ package.json
@@ -12,10 +12,11 @@
   },
   "keywords": [
     "visual regression",
+    "image comparison",
+    "wdio",
     "webdriverio",
     "wdio-service",
-    "wdio-plugin",
-    "comparison"
+    "wdio-plugin"
   ],
   "repository": {
     "type": "git",
@@ -31,18 +32,18 @@
     "resemblejs": "^3.2.4"
   },
   "devDependencies": {
-    "@types/jasmine": "^3.5.10",
-    "@types/node": "^13.13.2",
-    "@typescript-eslint/eslint-plugin": "^2.29.0",
-    "@typescript-eslint/parser": "^2.29.0",
+    "@types/jasmine": "^3.5.11",
+    "@types/node": "^14.0.14",
+    "@typescript-eslint/eslint-plugin": "^3.5.0",
+    "@typescript-eslint/parser": "^3.5.0",
     "@wdio/cli": "^6.1.2",
     "@wdio/jasmine-framework": "^6.1.0",
     "@wdio/local-runner": "^6.1.2",
     "@wdio/selenium-standalone-service": "^6.0.16",
     "@wdio/spec-reporter": "^6.0.16",
-    "eslint": "^6.8.0",
-    "ts-node": "^8.9.0",
-    "typescript": "^3.8.3"
+    "eslint": "^7.4.0",
+    "ts-node": "^8.10.2",
+    "typescript": "^3.9.6"
   },
   "files": [
     "build/**/*"
