+++ packages/internal.ts
@@ -1,5 +1,4 @@
 // This file avoid nesty circular dependency imports
-export * from './reporter';
 export * from './service/core/base-matcher';
 export * from './service/core/define-matcher';
 export * from './service/core/match';
@@ -7,5 +6,6 @@ export * from './service/matchers/element';
 export * from './service/matchers/viewport';
 export * from './service/browser-helpers';
 export * from './service';
+export * from './report';
 export * from './config';
 export * from './utils';
