--- packages/reporter/interfaces.d.ts
@@ -1,18 +1,0 @@
-interface MatchCommandResult {
-  fileName: string;
-  mismatch: number;
-  files: {
-    actual?: string;
-    expected?: string;
-    diff?: string;
-  };
-}
-
-export type TestContextResult = Omit<ReportData, 'matchers'>;
-
-export interface ReportData {
-  testName: string;
-  passed: boolean;
-  browser: string;
-  matchers: MatchCommandResult[];
-}
